Conflicts:
	LICENCE
	Makefile.doc
	jni/blackbox.c
	jni/bridges.c
	jni/cube.c
	jni/dominosa.c
	jni/fifteen.c
	jni/filling.c
	jni/flip.c
	jni/galaxies.c
	jni/guess.c
	jni/inertia.c
	jni/keen.c
	jni/lightup.c
	jni/loopy.c
	jni/magnets.c
	jni/map.c
	jni/midend.c
	jni/mines.c
	jni/net.c
	jni/netslide.c
	jni/nullgame.c
	jni/pattern.c
	jni/pearl.c
	jni/pegs.c
	jni/puzzles.h
	jni/rect.c
	jni/samegame.c
	jni/signpost.c
	jni/singles.c
	jni/sixteen.c
	jni/slant.c
	jni/solo.c
	jni/tents.c
	jni/towers.c
	jni/twiddle.c
	jni/unequal.c
	jni/untangle.c
	puzzles.but

+++ unfinished/numgame.c
@@ -1193,7 +1193,7 @@ int main(int argc, char **argv)
 	} else {
 	    if (nnumbers >= lenof(numbers)) {
 		fprintf(stderr, "%s: internal limit of %d numbers exceeded\n",
-			pname, lenof(numbers));
+			pname, (int)lenof(numbers));
 		return 1;
 	    } else {
 		numbers[nnumbers++] = atoi(p);
